+++ src/index.ts
@@ -1,15 +1,9 @@
 export {
-  EntityChangeFailureRateCard,
-  EntityDORAMetricsContent,
-  EntityDXDashboardContent,
-  EntityDeploymentFrequencyCard,
-  EntityOpenToDeployCard,
-  EntityLeadTimeCard,
-  EntityTimeToRecoveryCard,
-  EntityTopContributorsTable,
   EntityScorecardsCard,
   EntityTasksCard,
   EntityScorecardsPage,
   EntityTasksPage,
   dxPlugin,
 } from "./plugin";
+
+export { DxDataChartCard } from "./components/DxDataChartCard";