+++ src/plugin.ts
@@ -3,7 +3,6 @@ import {
   createApiFactory,
   createComponentExtension,
   createPlugin,
-  createRoutableExtension,
   discoveryApiRef,
   fetchApiRef,
 } from "@backstage/core-plugin-api";
@@ -29,124 +28,6 @@ export const dxPlugin = createPlugin({
     }),
   ],
 });
-
-/**
- * @deprecated
- */
-export const EntityDXDashboardContent = dxPlugin.provide(
-  createRoutableExtension({
-    name: "EntityDXDashboardContent",
-    component: () =>
-      import("./components/EntityDXDashboardContent").then(
-        (m) => m.EntityDXDashboardContent,
-      ),
-    mountPoint: rootRouteRef,
-  }),
-);
-
-/**
- * @deprecated
- */
-export const EntityDORAMetricsContent = dxPlugin.provide(
-  createRoutableExtension({
-    name: "EntityDORAMetricsContent",
-    component: () =>
-      import("./components/EntityDORAMetricsContent").then(
-        (m) => m.EntityDORAMetricsContent,
-      ),
-    mountPoint: rootRouteRef,
-  }),
-);
-
-/**
- * @deprecated
- */
-export const EntityChangeFailureRateCard = dxPlugin.provide(
-  createComponentExtension({
-    name: "EntityChangeFailureRateCard",
-    component: {
-      lazy: () =>
-        import("./components/EntityChangeFailureRateCard").then(
-          (m) => m.EntityChangeFailureRateCard,
-        ),
-    },
-  }),
-);
-
-/**
- * @deprecated
- */
-export const EntityDeploymentFrequencyCard = dxPlugin.provide(
-  createComponentExtension({
-    name: "EntityDeploymentFrequencyCard",
-    component: {
-      lazy: () =>
-        import("./components/EntityDeploymentFrequencyCard").then(
-          (m) => m.EntityDeploymentFrequencyCard,
-        ),
-    },
-  }),
-);
-
-/**
- * @deprecated
- */
-export const EntityOpenToDeployCard = dxPlugin.provide(
-  createComponentExtension({
-    name: "EntityOpenToDeployCard",
-    component: {
-      lazy: () =>
-        import("./components/EntityOpenToDeployCard").then(
-          (m) => m.EntityOpenToDeployCard,
-        ),
-    },
-  }),
-);
-
-/**
- * @deprecated
- */
-export const EntityLeadTimeCard = dxPlugin.provide(
-  createComponentExtension({
-    name: "EntityLeadTimeCard",
-    component: {
-      lazy: () =>
-        import("./components/EntityOpenToDeployCard").then(
-          (m) => m.EntityOpenToDeployCard,
-        ),
-    },
-  }),
-);
-
-/**
- * @deprecated
- */
-export const EntityTimeToRecoveryCard = dxPlugin.provide(
-  createComponentExtension({
-    name: "EntityTimeToRecoveryCard",
-    component: {
-      lazy: () =>
-        import("./components/EntityTimeToRecoveryCard").then(
-          (m) => m.EntityTimeToRecoveryCard,
-        ),
-    },
-  }),
-);
-
-/**
- * @deprecated
- */
-export const EntityTopContributorsTable = dxPlugin.provide(
-  createComponentExtension({
-    name: "EntityTopContributorsTable",
-    component: {
-      lazy: () =>
-        import("./components/EntityTopContributorsTable").then(
-          (m) => m.EntityTopContributorsTable,
-        ),
-    },
-  }),
-);
 
 export const EntityScorecardsCard = dxPlugin.provide(
   createComponentExtension({