+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 1.0.0 - 2025-07-10
+
+### Added
+
+- This adds a new `DxDataChart` component that can be used to display data from DX data studio queries!
+
+### Breaking changes
+
+- This version removes the following components which are made redundant by the `DxDataChart` flexibility.
+
+  - `EntityChangeFailureRateCard`
+  - `EntityDeploymentFrequencyCard`
+  - `EntityDORAMetricsContent`
+  - `EntityDXDashboardContent`
+  - `EntityLeadTimeCard`
+  - `EntityOpenToDeployCard`
+  - `EntityTimeToRecoveryCard`
+  - `EntityTopContributorsTable`
+
 ## 0.7.5 - 2025-06-13
 
 ### Added